requirements.txt
Django==4.2
django-admin-sortable2==2.1.10
django-filer==3.1.1
python-dotenv==1.0.0
mysqlclient==2.2.0
flake8==6.0.0